pyproject.toml
[project]
name = "ph-politician-rag"
version = "0.1.0"
description = "RAG pipeline for Philippine government documents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-pinecone>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "pinecone>=5.0.0,<7.0.0",
    "openai>=1.60.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]
